Xdim and Ydim should by int instead og size_tbecause in FFT_IDX2DIGFREQ_FAST it is substracted from another number and the results is a negative number

+++ software/em/xmipp/libraries/reconstruction/movie_filter_dose.h
@@ -138,7 +138,7 @@ public:
 
    /// Apply a dose filter to the image Fourier transform
    void applyDoseFilterToImage(
-		size_t Ydim, size_t Xdim,
+		int Ydim, int Xdim,
    		const MultidimArray< std::complex<double> > &FFT1,
 		const double dose_start, const double dose_finish
    		);
